src/test/java/com/astro/service/impl/ProductServiceImplTest.java
package com.astro.service.impl;

import com.astro.dataobject.ProductInfo;
import com.astro.enums.ProductStatusEnum;
import org.junit.Assert;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.data.domain.Page;
import org.springframework.data.domain.PageRequest;
import org.springframework.test.context.junit4.SpringRunner;

import java.math.BigDecimal;
import java.util.List;

import static org.junit.Assert.*;

/**
 * Created by astro on 2017/10/24.
 */
@SpringBootTest
@RunWith(SpringRunner.class)
public class ProductServiceImplTest {

    @Autowired
    private ProductServiceImpl productService;

    @Test
    public void findOne() throws Exception {
        ProductInfo one = productService.findOne("156");
        System.out.println(one);
        Assert.assertEquals("156",one.getProductId());
    }

    @Test
    public void findUpAll() throws Exception {
        List<ProductInfo> list = productService.findUpAll();
        Assert.assertNotEquals(0,list.size());
    }

    @Test
    public void findAll() throws Exception {
        PageRequest pageRequest = new PageRequest(0,2);
        Page<ProductInfo> list = productService.findAll(pageRequest);
        System.out.println(list.getTotalElements());


    }

    @Test
    public void save() throws Exception {
        ProductInfo productInfo=new ProductInfo();
        productInfo.setProductId("112");
        productInfo.setProductName("皮皮虾");
        productInfo.setProductPrice(new BigDecimal(3.2));
        productInfo.setProductStock(100);
        productInfo.setProductDescription("好吃");
        productInfo.setProductIcon("http:///1233453");
        productInfo.setProductStatus(0);
        productInfo.setCategoryType(1);

        ProductInfo save = productService.save(productInfo);
        Assert.assertNotNull(save);
    }

    @Test
    public void onSale() throws Exception{
        //ProductInfo productInfo = productService.findOne("1");
        ProductInfo productInfo = productService.onSale("1");
        Assert.assertEquals(ProductStatusEnum.UP.getCode(),productInfo.getProductStatus());
    }



}